requirements.txt
alibabacloud_dysmsapi20170525==3.1.2
alibabacloud_tea_openapi>=0.3.7
alibabacloud_tea_util>=0.3.8
flask==2.3.3
python-dotenv==1.0.0
werkzeug==2.3.7